backend/requirements.txt
# FastAPI backend
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Vector DB + LLM
qdrant-client==1.8.0
groq==0.28.0

# PDF + OCR
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0 

# Tokenizer for input limiting
tiktoken==0.6.0

# Utility
regex

# Embeddings (local)
sentence-transformers==2.2.2
